:bug: fix: corrigindo o tipo de pk do bibliotecario

+++ src/main/java/com/grupo04/Biblioteca/repository/BibliotecarioRepository.java
@@ -3,4 +3,4 @@ package com.grupo04.Biblioteca.repository;
 import com.grupo04.Biblioteca.models.BibliotecarioModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BibliotecarioRepository extends JpaRepository<BibliotecarioModel, Integer> { }
+public interface BibliotecarioRepository extends JpaRepository<BibliotecarioModel, Long> { }
